Core v0.20: fixes and improvements

+++ dash-spv-masternode-processor/src/types/coinbase_transaction.rs
@@ -9,9 +9,9 @@ pub struct CoinbaseTransaction {
     pub height: u32,
     pub merkle_root_mn_list: *mut [u8; 32],
     pub merkle_root_llmq_list: *mut [u8; 32],
-    pub best_cl_height_diff: u32,
+    pub best_cl_height_diff: u64,
     pub best_cl_signature: *mut [u8; 96],
-    pub credit_pool_balance: u64,
+    pub credit_pool_balance: i64,
 }
 
 impl Drop for CoinbaseTransaction {
